+++ tests/monotouch-test/ObjCRuntime/RuntimeTest.cs
@@ -481,7 +481,7 @@ namespace MonoTouchFixtures.ObjCRuntime {
 			while ((long) obj.RetainCount > (long) count / 2 && iterations++ < max_iterations) {
 				Thread.Sleep (100);
 			}
-			Assert.That (obj.RetainCount, Is.Not.GreaterThan (count / 2), "RC. Iterations: " + iterations);
+			Assert.That (obj.RetainCount, Is.Not.GreaterThan ((nuint) (count / 2)), "RC. Iterations: " + iterations);
 
 			obj.Dispose ();
 		}
