+++ tests/monotouch-test/Security/KeyTest.cs
@@ -129,7 +129,7 @@ namespace MonoTouchFixtures.Security {
 #endif
 
 					using (var attrs = public_key.GetAttributes ()) {
-						Assert.That (attrs.Count, Is.GreaterThan (0), "public/GetAttributes");
+						Assert.That (attrs.Count, Is.GreaterThan ((nuint) 0), "public/GetAttributes");
 					}
 					using (var data = public_key.GetExternalRepresentation (out error)) {
 						Assert.Null (error, "public/error-1");
@@ -158,7 +158,7 @@ namespace MonoTouchFixtures.Security {
 					}
 #endif
 					using (var attrs = private_key.GetAttributes ()) {
-						Assert.That (attrs.Count, Is.GreaterThan (0), "private/GetAttributes");
+						Assert.That (attrs.Count, Is.GreaterThan ((nuint) 0), "private/GetAttributes");
 					}
 					using (var data2 = private_key.GetExternalRepresentation (out error)) {
 						Assert.Null (error, "private/error-1");
@@ -505,7 +505,7 @@ namespace MonoTouchFixtures.Security {
 				using (var pub = key.GetPublicKey ())
 				using (var ex = key.GetKeyExchangeResult (SecKeyAlgorithm.EcdhKeyExchangeStandardX963Sha512, pub, p.Dictionary, out error)) {
 					Assert.Null (error, "GetKeyExchangeResult/error");
-					Assert.That (ex.Length, Is.EqualTo (p.RequestedSize), "GetKeyExchangeResult/result");
+					Assert.That (ex.Length, Is.EqualTo ((nuint) p.RequestedSize), "GetKeyExchangeResult/result");
 				}
 			}
 		}
